cdk/pyproject.toml
[tool.poetry]
name = "latency-based-routing-cdk"
version = "0.1.0"
description = "This is a simple project showing how to whole infrastructure with latency based routing using Amazon Route 53."
authors = ["sz3jdii <adam.swiatkowski@iodi.pl>"]
packages = [{ include = "latency_based_routing" }]

[tool.poetry.dependencies]
python = "^3.12"

# 3rdparties
aws-cdk-lib = ">=2.0.0"
constructs = ">=10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"